CSPEx v0.1.0 CSP.Channel

Module used to create and manage Channels

Options

There are some options that may be used to change a channel behavior, but the channel's options can only be set during it's creation.

The available options are:

name - Registers the channel proccess with a name. Note that the naming constraints are the same applied to a GenServer.

buffer_type - The type of the buffer used in the channel (by default :blocking).

buffer_size - The maximum capacity of the channel's buffer (by default 0).

Buffer types

There are three suported buffer types:

:blocking - The channel never blocks until the buffer capacity is full.

:sliding - The channel never blocks, but when the buffer is full, it starts discarding the older values on it to make room for the newer ones.

:dropping - The channel never blocks, but when the buffer is full, it starts discarding any new values that are put on it, keeping the old ones.

Collections interoperability

You can use a channel just like any collection:

channel = Channel.new
other_channel = Channel.new

spawn fn -> Enum.into(channel, other_channel) end

Channel.put(channel, "some_data")
Channel.get(other_channel) #=> "some_data"

All functions from Enum and Stream are available, but you must take into consideration the blocking operations:

channel = Channel.new

Enum.into(1..10, channel) # This line will block until someone reads the
ten values.

Example

Types

buffer_type :: :blocking | :sliding | :dropping

channel_ref :: term | t

option ::
  {:buffer_size, non_neg_integer} |
  {:buffer_type, buffer_type} |
  {:name, GenServer.name}

options :: [option]

t :: %CSP.Channel{pid: term}

Functions

close(channel)

Specs

close(channel_ref) :: :ok

Function responsible for closing a channel.

closed?(channel)

Specs

closed?(channel_ref) :: boolean

Returns true or false wheter the channel is closed.

get(channel)

Specs

get(channel_ref) :: term

Function responsible for fetching a value of the channel.

It will block until a value is inserted in the channel or it is closed.

member?(channel, value)

Specs

member?(channel_ref, term) :: boolean

Returns true or false wheter the value is present on the channel.

new(options \\ [])

Specs

new(options) :: t

Function responsible for creating a new channel.

Useful for using channels outside of a supervision tree.

Example

iex> chann = CSP.Channel.new
iex> spawn_link(fn -> CSP.Channel.put(chann, :data) end)
iex> CSP.Channel.get(chann)
:data

put(channel, item)

Specs

put(channel_ref, term) :: :ok

Function responsible for putting a value in the channel.

It may block until a value is fetched deppending on the buffer type of the channel.

size(channel)

Specs

size(channel_ref) :: non_neg_integer

Returns the current size of the channel.

start(options \\ [])

Specs

start(options) :: GenServer.on_start

Non-linking version of CSP.Channel.start_link/1

start_link(options \\ [])

Specs

start_link(options) :: GenServer.on_start

Function responsible for the starting of the channel.

Ideal for using a CSP in a supervision tree.